Separated browser extension file back into separate platform-specific files
